+++ test/main.js
@@ -1,5 +1,5 @@
 import a from './down.svg';
-import b from './up.svg?dataUrl';
+import b from './up.svg?datauri';
 
 import jenkins from './jenkins.svg';